requirements.txt
alembic==1.8.1
anyio==3.6.1
apache-airflow==2.4.1
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.2.1
apispec==3.3.2
argcomplete==2.0.0
asttokens==2.0.8
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
bigquery==0.0.32
blinker==1.5
cachelib==0.9.0
cachetools==5.2.0
cattrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
cron-descriptor==1.2.31
croniter==1.3.7
cryptography==36.0.2
dacktool==0.0.7
dbstream==0.1.16
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
entrypoints==0.4
exceptiongroup==1.0.0rc9
executing==1.1.1
Flask==2.2.2
Flask-AppBuilder==4.1.3
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gcloud==0.18.3
google-api-core==2.10.2
google-api-python-client==1.7.11
google-auth==2.12.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.5.3
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-secret-manager==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
googleauthentication==0.0.16
graphviz==0.20.1
greenlet==1.1.3
grpc-google-iam-v1==0.12.4
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.4
inflection==0.5.1
ipykernel==6.16.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kaggle==1.5.12
lazy-object-proxy==1.7.1
libcst==0.4.7
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.1
mdurl==0.1.2
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.1
packaging==21.3
pandas==1.3.4
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prison==0.2.1
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.7
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-daemon==2.3.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.2.1
pytzdata==2020.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.27
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.3
stack-data==0.5.1
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.1.0
termcolor==2.0.1
text-unidecode==1.3
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing-inspect==0.8.0
typing_extensions==4.3.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
WTForms==2.3.3
